requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.2
Automat==20.2.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.1.4
django-crispy-forms==1.10.0
Fabric3==1.14.post1
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==1.0.2
paramiko==2.7.2
Pillow==8.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pytz==2020.4
redis==3.5.3
redis-server==6.0.9
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.26.2
zope.interface==5.2.0
